pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cursfig"
version = "0.1.0"
description = "Configuration backup and restore tool"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "textual>=0.50",
]

[project.optional-dependencies]
github = ["PyGithub>=2.0"]
gdrive = ["google-api-python-client>=2.0", "google-auth-oauthlib>=1.0"]
all = ["cursfig[github,gdrive]"]

[project.scripts]
cursfig = "cursfig.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["cursfig*"]

[tool.setuptools.package-data]
cursfig = ["data/*.yaml"]
